refactor(agent-core-v2): drop the llm state machine, let turn invoke the request actor directly
